+++ src/test/java/com/green/gogiro/butchershop/ButcherShopMapperTest.java
@@ -156,7 +156,7 @@ class ButcherShopMapperTest {
     vo.setY("35.1804742090837");
     vo.setStar(5);
     vo.setIsBook(0);
-    ButDto dto = new ButDto(0,1);
+    ButDto dto = new ButDto((long)0,1);
     ButcherShopDetailVo vo2 = mapper.selButcherShopDetail(dto);
     assertEquals(vo,vo2);
     }
